+++ src/main/java/de/blablubbabc/shopkeepers/bluemap/ShopkeepersBlueMapPermission.java
@@ -1,0 +1,19 @@
+package de.blablubbabc.shopkeepers.bluemap;
+
+/**
+ * The plugin permissions.
+ */
+public class ShopkeepersBlueMapPermission {
+
+	/**
+	 * The {@code shopkeepers-bluemap.help} permission.
+	 */
+	public static String HELP = "shopkeepers-bluemap.help";
+	/**
+	 * The {@code shopkeepers-bluemap.reload} permission.
+	 */
+	public static String RELOAD = "shopkeepers-bluemap.reload";
+
+	private ShopkeepersBlueMapPermission() {
+	}
+}